moved DualSource back to operations
