+++ oamap/util.py
@@ -271,101 +271,3 @@ def returntype(fcn, paramtypes):
             return None
         else:
             return overload.signature.return_type
-
-class DualSource(object):
-    def __init__(self, old, oldns):
-        self.old = old
-        self.new = {}
-
-        i = 0
-        self.namespace = None
-        while self.namespace is None or self.namespace in oldns:
-            self.namespace = "namespace-" + str(i)
-            i += 1
-
-        self._arraynum = 0
-
-    def arrayname(self):
-        trial = None
-        while trial is None or trial in self.new:
-            trial = "array-" + str(self._arraynum)
-            self._arraynum += 1
-        return trial
-
-    def getall(self, roles):
-        out = {}
-
-        if hasattr(self.old, "getall"):
-            out.update(self.old.getall([x for x in roles if x.namespace != self.namespace]))
-        else:
-            for x in roles:
-                if x.namespace != self.namespace:
-                    out[x] = self.old[str(x)]
-
-        if hasattr(self.new, "getall"):
-            out.update(self.new.getall([x for x in roles if x.namespace == self.namespace]))
-        else:
-            for x in roles:
-                if x.namespace == self.namespace:
-                    out[x] = self.new[str(x)]
-
-        return out
-
-    def put(self, schemanode, *arrays):
-        if isinstance(schemanode, oamap.schema.Primitive):
-            datarole = oamap.generator.DataRole(self.arrayname(), self.namespace)
-            roles2arrays = {datarole: arrays[0]}
-            schemanode.data = str(datarole)
-
-        elif isinstance(schemanode, oamap.schema.List):
-            startsrole = oamap.generator.StartsRole(self.arrayname(), self.namespace, None)
-            stopsrole = oamap.generator.StopsRole(self.arrayname(), self.namespace, None)
-            startsrole.stops = stopsrole
-            stopsrole.starts = startsrole
-            roles2arrays = {startsrole: arrays[0], stopsrole: arrays[1]}
-            schemanode.starts = str(startsrole)
-            schemanode.stops = str(stopsrole)
-
-        elif isinstance(schemanode, oamap.schema.Union):
-            tagsrole = oamap.generator.TagsRole(self.arrayname(), self.namespace, None)
-            offsetsrole = oamap.generator.OffsetsRole(self.arrayname(), self.namespace, None)
-            tagsrole.offsets = offsetsrole
-            offsetsrole.tags = tagsrole
-            roles2arrays = {tagsrole: arrays[0], offsetsrole: arrays[1]}
-            schemanode.tags = str(tagsrole)
-            schemanode.offsets = str(offsetsrole)
-
-        elif isinstance(schemanode, oamap.schema.Record):
-            pass
-
-        elif isinstance(schemanode, oamap.schema.Tuple):
-            pass
-
-        elif isinstance(schemanode, oamap.schema.Pointer):
-            positionsrole = oamap.generator.PositionsRole(self.arrayname(), self.namespace)
-            roles2arrays = {positionsrole: arrays[0]}
-            schemanode.positions = str(positionsrole)
-
-        else:
-            raise AssertionError(schemanode)
-
-        if schemanode.nullable:
-            maskrole = oamap.generator.MaskRole(self.arrayname(), self.namespace, roles2arrays)
-            roles2arrays = dict(list(roles2arrays.items()) + [(maskrole, arrays[-1])])
-            schemanode.mask = str(maskrole)
-
-        schemanode.namespace = self.namespace
-        self.putall(roles2arrays)
-
-    def putall(self, roles2arrays):
-        if hasattr(self.new, "putall"):
-            self.new.putall(roles2arrays)
-        else:
-            for n, x in roles2arrays.items():
-                self.new[str(n)] = x
-
-    def close(self):
-        if hasattr(self.old, "close"):
-            self.old.close()
-        if hasattr(self.new, "close"):
-            self.new.close()
